fix compile error in TcgPei

git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@13699 6f19259b-4bc3-4df7-8a09-765794883524

+++ SecurityPkg/Tcg/TcgPei/TcgPei.c
@@ -185,7 +185,7 @@ MeasureCRTMVersion (
 
   TcgEventHdr.PCRIndex  = 0;
   TcgEventHdr.EventType = EV_S_CRTM_VERSION;
-  TcgEventHdr.EventSize = StrSize((CHAR16*)PcdGetPtr (PcdFirmwareVersionString));
+  TcgEventHdr.EventSize = (UINT32) StrSize((CHAR16*)PcdGetPtr (PcdFirmwareVersionString));
 
   return HashLogExtendEvent (
            PeiServices,
